backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0  # PostgreSQL driver
alembic==1.13.3  # Migrations
pydantic-settings==2.5.2  # Env config
pyjwt[crypto]==2.9.0  # JWT
passlib[argon2]==1.7.4  # Password hashing
cryptography==43.0.1  # Fernet AES-256
httpx==0.27.0  # HTTP client for OAuth
python-multipart==0.0.9  # Form data
pytest==8.3.3
pytest-asyncio==0.24.0
responses==0.25.0  # Mock APIs for tests
celery==5.4.0  # Prep for Phase 3
redis==5.2.1